refactor: moved intrs to numeric dir
