+++ instructions/numeric/i32_bitop.cpp
