+++ instructions/numeric/i32_bitop.hpp
