+++ instructions/numeric/i32_comparison.cpp
