+++ instructions/numeric/i32_comparison.hpp
