+++ instructions/numeric/i64_bitop.cpp
