+++ instructions/numeric/i64_bitop.hpp
