+++ scanner/instruction_parser.hpp
@@ -1,16 +1,6 @@
 #ifndef ANTIWASM_INSTRUCTION_PARSER_HPP
 #define ANTIWASM_INSTRUCTION_PARSER_HPP
 
-#include "arithmetic/f32_arithmetic.hpp"
-#include "arithmetic/f64_arithmetic.hpp"
-#include "arithmetic/i32_arithmetic.hpp"
-#include "arithmetic/i64_arithmetic.hpp"
-#include "bitop/i32_bitop.hpp"
-#include "bitop/i64_bitop.hpp"
-#include "comparison/f32_comparison.hpp"
-#include "comparison/f64_comparison.hpp"
-#include "comparison/i32_comparison.hpp"
-#include "comparison/i64_comparison.hpp"
 #include "const/f32_const.hpp"
 #include "const/f64_const.hpp"
 #include "const/i32_const.hpp"
@@ -26,6 +16,16 @@
 #include "memory/memory_size.hpp"
 #include "memory/store.hpp"
 #include "nop.hpp"
+#include "numeric/f32_arithmetic.hpp"
+#include "numeric/f32_comparison.hpp"
+#include "numeric/f64_arithmetic.hpp"
+#include "numeric/f64_comparison.hpp"
+#include "numeric/i32_arithmetic.hpp"
+#include "numeric/i32_bitop.hpp"
+#include "numeric/i32_comparison.hpp"
+#include "numeric/i64_arithmetic.hpp"
+#include "numeric/i64_bitop.hpp"
+#include "numeric/i64_comparison.hpp"
 #include "unreachable.hpp"
 #include "variable/global_get.hpp"
 #include "variable/global_set.hpp"
