+++ test/scanner/instruction_parser_test.cpp
@@ -1,13 +1,5 @@
 #define BOOST_TEST_DYN_LINK
 
-#include "arithmetic/f32_arithmetic.cpp"
-#include "arithmetic/f64_arithmetic.cpp"
-#include "arithmetic/i32_arithmetic.cpp"
-#include "arithmetic/i64_arithmetic.cpp"
-#include "comparison/f32_comparison.cpp"
-#include "comparison/f64_comparison.cpp"
-#include "comparison/i32_comparison.cpp"
-#include "comparison/i64_comparison.cpp"
 #include "const/f32_const.cpp"
 #include "const/f64_const.cpp"
 #include "const/i32_const.cpp"
@@ -22,9 +14,17 @@
 #include "memory/memory_init.cpp"
 #include "memory/memory_size.cpp"
 #include "memory/store.cpp"
-#include "bitop/i32_bitop.cpp"
-#include "bitop/i64_bitop.cpp"
 #include "nop.cpp"
+#include "numeric/f32_arithmetic.cpp"
+#include "numeric/f32_comparison.cpp"
+#include "numeric/f64_arithmetic.cpp"
+#include "numeric/f64_comparison.cpp"
+#include "numeric/i32_arithmetic.cpp"
+#include "numeric/i32_bitop.cpp"
+#include "numeric/i32_comparison.cpp"
+#include "numeric/i64_arithmetic.cpp"
+#include "numeric/i64_bitop.cpp"
+#include "numeric/i64_comparison.cpp"
 #include "unreachable.cpp"
 #include "variable/global_get.cpp"
 #include "variable/global_set.cpp"
